backend-flask/app/main/controller/auth_controller.py
from flask import request
from flask_restx import Resource

from app.main.service.auth_service import Auth
from ..util.dto import AuthDto

api = AuthDto.api
user_auth = AuthDto.user_auth


@api.route('/login/')
class UserLogin(Resource):
    """ User login res """
    @api.doc('user login')
    @api.expect(user_auth, validate=True)
    def post(self):
        post_data = request.json
        return Auth.login_user(data=post_data)
    

@api.route('/logout/')
class UserLogout(Resource):
    """ User Logout res """
    @api.doc('logout a user')
    def post(self):
        # get auth token
        auth_header = request.headers.get('Authorization')
        return Auth.logout_user(data=auth_header)
