backend-flask/requirements.txt
alembic==1.5.5
aniso8601==9.0.0
appdirs==1.4.4
attrs==20.3.0
autopep8==1.5.6
bcrypt==3.2.0
bidict==0.21.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
dataclasses==0.8
distlib==0.3.1
dnspython==1.16.0
eventlet==0.30.1
filelock==3.0.12
flake8==3.9.0
Flask==2.0.3
Flask-API==2.0
Flask-Bcrypt==0.7.1
Flask-Migrate==2.7.0
flask-redis==0.4.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
gevent==21.1.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.53.0
greenlet==1.0.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
importlib-metadata==3.7.0
importlib-resources==5.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
jwt==1.2.0
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
oauth2client==4.1.3
packaging==20.9
pkg_resources==0.0.0
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==2.0.3
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
